Add water intake summary and bar graph visualization; implement week start calculation

+++ lib/widgets/bargraph/individual_bar.dart
@@ -1,9 +1,9 @@
 class IndividualBar {
-  final double x;
+  final int x;
   final double y;
 
   IndividualBar({
     required this.x,
     required this.y,
   }); //y is the height = amount water
-}
+}
